Merge commit from fork
* chore:Fix a meson error.

* Fix potential security issue due to unknown key types.

HS256, HS384 & HS512 are symmetric key based algorithms that use an
array of bytes as the key.  The array of bytes could be the public key
of a RSxxx, ECxxx or PSxxx algorithm.  This means that unless the user
of this library validated that the jwt algorithm matched the expected
algorithm type(s), the user could be using a compromised JWT and not
realize it.

This fix requires the caller to declare if they expect a symmetric
cypher or not to prevent this attack and maintain the interface.  With
this change pub/private algorithms and symmetric algorithms are
segmented to protect the user from this kind of attack.